Move ColorInput into the inputs folder
