+++ frontend/src/components/widgets/inputs/ColorInput.svelte
